outlined remaining test files before reconfiguring install scripts
