+++ oauth2_gin_test.go
@@ -1,0 +1,25 @@
+package main
+
+import (
+	"testing"
+)
+
+func TestOAuthGoogleCallback(t *testing.T) {
+
+}
+
+func TestOAuthGoogleLogin(t *testing.T) {
+
+}
+
+func TestOAuthGoogleOAuthn(t *testing.T) {
+
+}
+
+func TestGenerateStateOauthCookie(t *testing.T) {
+
+}
+
+func TestGetUserDataFromGoogle(t *testing.T) {
+
+}
