+++ object_converter_test.go
@@ -1,0 +1,93 @@
+package main
+
+import (
+	"testing"
+)
+
+func TestCompileShopifyToSystemProduct(t *testing.T) {
+
+}
+
+func TestCompileRemoveQueueFilter(t *testing.T) {
+
+}
+
+func TestConvertDatabaseToRegister(t *testing.T) {
+
+}
+
+func TestConvertDatabaseToWarehouse(t *testing.T) {
+
+}
+
+func TestCompileQueueFilterSearch(t *testing.T) {
+
+}
+
+func TestConvertProductToShopify(t *testing.T) {
+
+}
+
+func TestConvertVariantToShopifyProdVariant(t *testing.T) {
+
+}
+
+func TestConvertToShopifyIDs(t *testing.T) {
+
+}
+
+func TestConvertVariantToShopify(t *testing.T) {
+
+}
+
+func TestConvertVariantToShopifyVariant(t *testing.T) {
+
+}
+
+func TestCompileShopifyOptions(t *testing.T) {
+
+}
+
+func TestCompileCustomerSearchData(t *testing.T) {
+
+}
+
+func TestCompileCustomerData(t *testing.T) {
+
+}
+
+func TestCompileOrderSearchResult(t *testing.T) {
+
+}
+
+func TestCompileOrderData(t *testing.T) {
+
+}
+
+func TestCompileFilterSearch(t *testing.T) {
+
+}
+
+func TestCompileProductImages(t *testing.T) {
+
+}
+
+func TestCompileSearchResult(t *testing.T) {
+
+}
+
+func TestConvertProductToCSVProduct(t *testing.T) {
+
+}
+
+func TestCompileProduct(t *testing.T) {
+
+}
+
+func TestCompileVariants(t *testing.T) {
+
+}
+
+func TestCompileVariantByID(t *testing.T) {
+
+}
