+++ restrictions_test.go
@@ -1,0 +1,33 @@
+package main
+
+import (
+	"testing"
+)
+
+func TestFetchRestrictionsToMap(t *testing.T) {
+
+}
+
+func TestPushRestrictionsToMap(t *testing.T) {
+
+}
+
+func TestApplyFetchRestriction(t *testing.T) {
+
+}
+
+func TestDeterPushRestriction(t *testing.T) {
+
+}
+
+func TestApplyPushRestrictionProduct(t *testing.T) {
+
+}
+
+func TestApplyPushRestrictionVt(t *testing.T) {
+
+}
+
+func TestDeterFetchRestriction(t *testing.T) {
+
+}
